nf: Kotlin Cleanup (#12193)
* refactor: Kotlin Cleanup - ACRATest

* refactor: Kotlin Cleanup - DeckPickerImportTest

* refactor: Kotlin Cleanup - AnkiAssert

* refactor: Kotlin Cleanup - AnkiActivityUtils
