+++ AnkiDroid/src/test/java/com/ichi2/anki/DeckPickerImportTest.kt
@@ -19,7 +19,6 @@ import android.content.Intent
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.dialogs.AsyncDialogFragment
 import com.ichi2.anki.dialogs.ImportDialog
-import com.ichi2.utils.KotlinCleanup
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers
 import org.junit.Assert.fail
@@ -48,7 +47,6 @@ class DeckPickerImportTest : RobolectricTest() {
 
     private class DeckPickerImport : DeckPicker() {
         private var mDialogFragment: AsyncDialogFragment? = null
-        @KotlinCleanup("invert if and return early")
         fun getAsyncDialogFragmentClass(): Class<*> {
             if (mDialogFragment == null) {
                 fail("No async fragment shown")
